# Speed up `brew list`